FiFoBuffer: missed synchronize added. LogEventData.id is now of type Long.

+++ projectforge-common/src/main/kotlin/org/projectforge/common/logging/LogFilter.kt
@@ -43,5 +43,5 @@ class LogFilter(
   /**
    * @return If given, all log entries with order orderNumber higher than this orderNumber will be queried.
    */
-  var lastReceivedLogOrderNumber: Int? = null
+  var lastReceivedLogOrderNumber: Long? = null
 )
